Tighten SSH option passthrough
Previously, options without values (e.g. `-t`) would be passed as:
('-t', '')

When passed through to the shell, this became the following:
-t ''

This is unfortunately not the same as intended:
-t

Additionally, previously we indiscriminately appended `--` to the
`az ssh vm` command, even if it was not needed. Now, we only do that
when options or arguments are specified.